DSH1-B: move the DSH web process core into ccteam-harness
The per-identity `dsh web` runtime moves out of `ccteam-web::dsh_web` into
`ccteam-harness::execution::dsh_runtime::DshRuntimeManager`. The DSH adapter
cannot depend on ccteam-web (ccteam-web -> ccteam-im -> ccteam-harness), so
"one identity, one runtime" now holds because BOTH consumers share one
instance, not because each honors a convention: `ccteam start` builds ONE
manager in its composition root and hands it to web serve, which configures
it (enabled / daemon_url / attach_url) once the bind is known. Standalone
`ccteam web` builds its own; an unconfigured manager answers `disabled` and
spawns nothing.

Moved verbatim where possible: instance map keyed by owner tag, start-claim
dedup + inflight orphan healing, attach-if-detected (never double-writing a
home), spawn-until-readiness + PDEATHSIG + health probe, stderr tail ring,
graceful stop, shutdown_all, the legacy orphan sweep. The locking discipline
is preserved exactly — no MutexGuard is held across an `.await` that
re-locks (the production self-deadlock its regression test guards).

`ccteam-web` keeps everything web: the companion-port byte proxy with HTML
buffering + the secure-context `crypto.randomUUID` polyfill, the REST shapes
(`DshStatusResponse` / `DshWebStatusState` / `DshHomeKind`, unchanged JSON),
the companion port, the operator-only `native_url` gate, and the
`Identity` -> `DshRuntimeIdentity` mapping. Enrollment bearers are injected
as `DshEnrollmentResolver` because `ccteam-core::enroll` sits above harness.

Pain point: requirements.md "one DSH process per identity" / tech-design
§ DSH web companion. No REST, SPA, attach, seeding or tenant-home change.

cargo test -p ccteam-web --test dsh_web_test: 8/8 (deadlock + cancelled-start
regressions included). harness --lib 511, web --lib 172, test-baseline 1951
passed (only the pre-existing ccteam-im gateway resume flake, reproduced 3/3
on a pristine tree). clippy 0 warnings, fmt clean.
