+++ crates/ccteam-harness/src/lib.rs
@@ -75,6 +75,10 @@ pub use execution::dsh_acp::{
     resolve_dsh_default_bin, tenant_home_segment, DshAcpAdapter, DshConfigSource,
     DshWebSpawnOptions, DSH_ACP_ADAPTER_NAME, DSH_BIN_ENV, DSH_NATIVE_WEB_PROFILE, DSH_WEB_PROFILE,
 };
+pub use execution::dsh_runtime::{
+    sweep_legacy_dsh_orphans, DshEnrollmentResolver, DshRuntimeConfig, DshRuntimeIdentity,
+    DshRuntimeManager, DshRuntimeState, DshRuntimeStatus,
+};
 pub use execution::fs_atomic::atomic_write_durable;
 pub use execution::grok_acp::{GrokAcpAdapter, GROK_ACP_ADAPTER_NAME};
 pub use execution::host_channel::{
